requirements.txt
streamlit
pandas
plotly
requests
google-genai
beautifulsoup4
lxml